requirements.txt
sentence_transformers
streamlit
pinecone_client
pypdf